+++ src/camera.h
@@ -9,12 +9,12 @@ typedef struct {
     Vec3 x, y, z;
     Vec3 lookAt;
     uint32_t width, height;
-    double FOV, aspectRatio;
+    float FOV, aspectRatio;
     Vec3 renderTargetCenter;
-    double renderTargetWidth, renderTargetHeight, renderTargetDistance;
+    float renderTargetWidth, renderTargetHeight, renderTargetDistance;
 } Camera;
 
-Camera* camera_create(Vec3 position, Vec3 lookAt, uint32_t width, uint32_t height, double FOV);
+Camera* camera_create(Vec3 position, Vec3 lookAt, uint32_t width, uint32_t height, float FOV);
 void camera_destroy(Camera* camera);
 
 #endif //RAYTRACER_CAMERA_H
