+++ src/utils/vec3.h
@@ -9,24 +9,24 @@
 
 typedef union {
     struct {
-        double x, y, z;
+        float x, y, z;
     };
     struct {
-        double r, g, b;
+        float r, g, b;
     };
 } Vec3;
 
 Vec3 vec3_add(Vec3 a, Vec3 b);
 Vec3 vec3_sub(Vec3 a, Vec3 b);
-double vec3_dot(Vec3 a, Vec3 b);
+float vec3_dot(Vec3 a, Vec3 b);
 Vec3 vec3_cross(Vec3 a, Vec3 b);
-Vec3 vec3_clamp(Vec3 a, double min, double max);
+Vec3 vec3_clamp(Vec3 a, float min, float max);
 Vec3 vec3_hadamard(Vec3 a, Vec3 b);
 Vec3 vec3_norm(Vec3 a);
-double vec3_length(Vec3 a);
-Vec3 vec3_offset(Vec3 a, double offset);
-Vec3 vec3_mul(Vec3 a, double b);
-Vec3 vec3_div(Vec3 a, double b);
+float vec3_length(Vec3 a);
+Vec3 vec3_offset(Vec3 a, float offset);
+Vec3 vec3_mul(Vec3 a, float b);
+Vec3 vec3_div(Vec3 a, float b);
 Vec3 vec3_reflect(Vec3 incomingVec, Vec3 normal);
 
 #endif //RAYTRACER_VEC3_H
